Allow configuration of Database 12.2.0.1 box time zone (#33)
This implements the enhancement in Issue #33. By default, the VM time zone is set to the host time zone. In the environment section of the Vagrantfile, a different time zone can be specified either by name (e.g., "America/Los_Angeles") or by an offset from GMT (e.g., "Etc/GMT-2").

Changes are:
- In the Vagrantfile, add code to determine host time zone
- In the Vagrantfile, add a variable in the environment section to pass the time zone to install.sh
- In install.sh, add code to set the VM's time zone using the environment variable passed in from the Vagrantfile
- Update README.md to reflect the enhancement

Signed-off-by: Paul Neumann <38871902+PaulNeumann@users.noreply.github.com>