Prevent EOL conversion for .sh scripts for DB boxes (#71)
This addresses issue #71 by preventing automatic end-of-line conversion from LF to CR/LF for shell scripts for the 3 Oracle Database boxes. By default, Git for Windows converts line endings from LF to CR/LF. (This default can be overridden with the core.autocrlf parameter.)

This commit adds the same .gitattributes file that @AmedeeBulle created for the ContainerRegistry and Kubernetes boxes to the 11.2.0.2, 12.2.0.1 and 18.3.0 database boxes. The .gitattributes file tells Git to retain the LF line endings for .sh files.

Signed-off-by: Paul Neumann <38871902+PaulNeumann@users.noreply.github.com>