Explicitly set 18c XE box hostname (#91)
For the Oracle Database 18c XE box, the hostname must be "localhost", or listener configuration will fail. Since "localhost" is the default hostname, the "config.vm.name = NAME" line was originally commented out in the Vagrantfile. However, if another Vagrantfile that specifies a hostname is merged with this box's Vagrantfile, the box may end up using that hostname instead of "localhost". (See https://www.vagrantup.com/docs/vagrantfile/#load-order-and-merging.) This seems to be what happened in Issue #91.

This change explicitly sets the hostname for the Oracle Database 18c XE box to "localhost" in the Vagrantfile, to reduce the chance of it being overridden by a hostname specified in another Vagrantfile.

Signed-off-by: Paul Neumann <38871902+PaulNeumann@users.noreply.github.com>