Update README.md files for #35
Update the README.md files for the single-instance database boxes to describe how to configure the boxes using environment variables or (optionally, but preferably) the vagrant-env plugin. The updated text and formatting are based on the OLCNE box README.md file.

Also add newlines around headings and lists to suppress markdownlint warnings

Signed-off-by: Paul Neumann <38871902+PaulNeumann@users.noreply.github.com>